simplify

+++ install.sh
@@ -4,16 +4,14 @@ THIS_DIR=$(cd $(dirname $0); pwd)
 
 #echo $THIS_DIR
 
-cd ~
+cd ${HOME}
 
 for file in .emacs.d .screenrc  .zshrc .gitconfig .minttyrc
 do
-  echo ln -s dotfiles/$file .
-  ln -s dotfiles/$file .
+  [ ! -e $file ] && ln -s dotfiles/$file .
 done
 
-echo dotfiles/.gitconfig.local.template ~/.gitconfig.local
-cp   dotfiles/.gitconfig.local.template ~/.gitconfig.local
+[ ! -e .gitconfig.local ] && cp dotfiles/.gitconfig.local.template .gitconfig.local
 
-echo "edit .gitconfig.local manually"
+echo "please edit .gitconfig.local manually"
